[Extensions c2s] Remove extension's request from UI after access granted
An extension's site access request is removed after the extension is
granted access. This CL notifies the observers of the request removal,
so the extensions UI (request access button and extensions menu) can
be updated.

Bug: 330588494
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5496311
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1294041}

+++ extensions/browser/permissions_manager.cc
@@ -804,7 +804,10 @@ void PermissionsManager::RemoveSiteAccessRequest(
     return;
   }
 
-  helper->RemoveRequest(extension_id);
+  bool request_removed = helper->RemoveRequest(extension_id);
+  if (!request_removed) {
+    return;
+  }
 
   if (!helper->HasRequests()) {
     DeleteSiteAccessRequestHelperFor(tab_id);
@@ -842,7 +845,15 @@ void PermissionsManager::NotifyExtensionPermissionsUpdated(
     UpdateReason reason) {
   std::vector<int> tabs_to_remove;
   for (auto& [tab_id, helper] : requests_helpers_) {
-    helper->RemoveRequestIfGrantedAccess(extension);
+    bool request_removed = helper->RemoveRequestIfGrantedAccess(extension);
+    if (!request_removed) {
+      continue;
+    }
+
+    for (auto& observer : observers_) {
+      observer.OnSiteAccessRequestRemoved(extension.id(), tab_id);
+    }
+
     if (!helper->HasRequests()) {
       tabs_to_remove.push_back(tab_id);
     }
